requirements.txt
# requirements.txt
torch
torchvision
kagglehub
pillow
numpy
pandas
scikit-learn
matplotlib
streamlit
